+++ include/apsis/geometry/rectangle.h
@@ -16,7 +16,7 @@ namespace Apsis {
       bool intersects(Rectangle* rectangle);
       bool intersects(Line* line);
 
-      unsigned int clip(Line* line, double* tMin, double* tMax);
+      unsigned int clip(Line* line, float* tMin, float* tMax);
 
       bool contains(Point* point);
