+++ src/geometry/line.cpp
@@ -6,11 +6,11 @@ void Apsis::Geometry::Line::unit(Apsis::Geometry::Line* line) {
   unit(line, 1.0);
 }
 
-void Apsis::Geometry::Line::unit(Apsis::Geometry::Line* line, double magnitude) {
-  double magnitudeX = points[1].x - points[0].x;
-  double magnitudeY = points[1].y - points[0].y;
+void Apsis::Geometry::Line::unit(Apsis::Geometry::Line* line, float magnitude) {
+  float magnitudeX = points[1].x - points[0].x;
+  float magnitudeY = points[1].y - points[0].y;
 
-  double magnitudeT = magnitudeX * magnitudeX + magnitudeY * magnitudeY;
+  float magnitudeT = magnitudeX * magnitudeX + magnitudeY * magnitudeY;
   magnitudeT = sqrt(magnitudeT);
 
   // Find unit magnitudes
@@ -23,34 +23,34 @@ void Apsis::Geometry::Line::unit(Apsis::Geometry::Line* line, double magnitude) 
   line->points[1].y = points[0].y + magnitudeY * magnitude;
 }
 
-double Apsis::Geometry::Line::magnitude() {
-  double magnitudeX = points[1].x - points[0].x;
-  double magnitudeY = points[1].y - points[0].y;
+float Apsis::Geometry::Line::magnitude() {
+  float magnitudeX = points[1].x - points[0].x;
+  float magnitudeY = points[1].y - points[0].y;
 
-  double magnitude = magnitudeX * magnitudeX + magnitudeY * magnitudeY;
+  float magnitude = magnitudeX * magnitudeX + magnitudeY * magnitudeY;
   magnitude = sqrt(magnitude);
 
   return magnitude;
 }
 
 bool Apsis::Geometry::Line::intersects(Line* line) {
-  double x1  = points[0].x;
-  double x2  = line->points[0].x;
+  float x1  = points[0].x;
+  float x2  = line->points[0].x;
 
-  double y1  = points[0].y;
-  double y2  = line->points[0].y;
+  float y1  = points[0].y;
+  float y2  = line->points[0].y;
 
-  double dX1 = points[1].x - points[0].x;
-  double dY1 = points[1].y - points[0].y;
+  float dX1 = points[1].x - points[0].x;
+  float dY1 = points[1].y - points[0].y;
 
-  double dX2 = line->points[1].x - line->points[0].x;
-  double dY2 = line->points[1].y - line->points[0].y;
+  float dX2 = line->points[1].x - line->points[0].x;
+  float dY2 = line->points[1].y - line->points[0].y;
 
   // Find t in (x1-x2)/(dX2-dX1) = t (for both tX and tY)
   // This formula is derived from the parametric equation for a line
 
-  double tX = (x1 - x2) / (dX2 - dX1);
-  double tY = (y1 - y2) / (dY2 - dY1);
+  float tX = (x1 - x2) / (dX2 - dX1);
+  float tY = (y1 - y2) / (dY2 - dY1);
 
   if (dX2 == dX1 && dY2 == dY1) {
     // Parallel
@@ -67,8 +67,8 @@ bool Apsis::Geometry::Line::intersects(Line* line) {
 }
 
 bool Apsis::Geometry::Line::clips(Line* line, Point* point) {
-  double x1, x2, x3, x4;
-  double y1, y2, y3, y4;
+  float x1, x2, x3, x4;
+  float y1, y2, y3, y4;
 
   x1 = this->points[0].x;
   x2 = this->points[1].x;
@@ -83,9 +83,9 @@ bool Apsis::Geometry::Line::clips(Line* line, Point* point) {
   // Find t in (x1-x2)/(dX2-dX1) = t (for both t1 and t2)
   // This formula is derived from the parametric equation for a line
 
-  double t1, t2;
+  float t1, t2;
 
-  double determinant = (y4 - y3) * (x2 - x1) - (x4 - x3) * (y2 - y1);
+  float determinant = (y4 - y3) * (x2 - x1) - (x4 - x3) * (y2 - y1);
 
   t1  = (x4 - x3) * (y1 - y3) - (y4 - y3) * (x1 - x3);
   t1 /= determinant;
